Request permission on app startup
