+++ app/src/main/java/ch/epfl/sdp/ui/main/swipe/SwipeFragment.java
@@ -2,6 +2,7 @@ package ch.epfl.sdp.ui.main.swipe;
 
 import android.annotation.SuppressLint;
 import android.app.AlertDialog;
+import android.location.Location;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -33,7 +34,9 @@ import ch.epfl.sdp.Event;
 import ch.epfl.sdp.R;
 import ch.epfl.sdp.databinding.FragmentSwipeBinding;
 import ch.epfl.sdp.db.Database;
+import ch.epfl.sdp.map.LocationService;
 import ch.epfl.sdp.platforms.firebase.db.FirestoreDatabase;
+import ch.epfl.sdp.platforms.google.map.GoogleLocationService;
 import ch.epfl.sdp.ui.main.MainActivity;
 
 public class SwipeFragment extends Fragment implements SwipeFlingAdapterView.onFlingListener {
@@ -46,7 +49,7 @@ public class SwipeFragment extends Fragment implements SwipeFlingAdapterView.onF
     private int mNumberSwipe = 0;
 
     private EventDetailFragment mInfoFragment;
-    private Event mCurrentEvent;
+    private LocationService mLocationService;
 
     @Override
     public void removeFirstObjectInAdapter() {
@@ -55,10 +58,14 @@ public class SwipeFragment extends Fragment implements SwipeFlingAdapterView.onF
     }
 
     @Override
-    public void onLeftCardExit(Object o) {mNumberSwipe +=1;}
+    public void onLeftCardExit(Object o) {
+        mNumberSwipe += 1;
+    }
 
     @Override
-    public void onRightCardExit(Object o) {mNumberSwipe +=1;}
+    public void onRightCardExit(Object o) {
+        mNumberSwipe += 1;
+    }
 
     @Override
     public void onAdapterAboutToEmpty(int i) {}
@@ -73,12 +80,14 @@ public class SwipeFragment extends Fragment implements SwipeFlingAdapterView.onF
     public SwipeFragment() {
         mFactory = new EventSwipeViewModel.EventSwipeViewModelFactory();
         mFactory.setDatabase(new FirestoreDatabase(FirebaseFirestore.getInstance()));
+        mLocationService = GoogleLocationService.getInstance();
     }
 
     @VisibleForTesting
-    public SwipeFragment(@NonNull Database database) {
+    public SwipeFragment(@NonNull Database database, @NonNull LocationService locationService) {
         mFactory = new EventSwipeViewModel.EventSwipeViewModelFactory();
         mFactory.setDatabase(database);
+        mLocationService = locationService;
     }
 
     @Override
@@ -86,7 +95,7 @@ public class SwipeFragment extends Fragment implements SwipeFlingAdapterView.onF
         mBinding = FragmentSwipeBinding.inflate(inflater, container, false);
         return mBinding.getRoot();
     }
-    
+
     @Override
     public void onActivityCreated(@Nullable Bundle savedInstanceState) {
         super.onActivityCreated(savedInstanceState);
@@ -106,8 +115,6 @@ public class SwipeFragment extends Fragment implements SwipeFlingAdapterView.onF
         dialog.positiveButton(null, "Choose", null);
         dialog.show();
 
-
-
         mEventList = new ArrayList<>();
         mArrayAdapter = new CardArrayAdapter(getContext(), mEventList);
         mBinding.cardsListView.setAdapter(mArrayAdapter);
@@ -123,12 +130,14 @@ public class SwipeFragment extends Fragment implements SwipeFlingAdapterView.onF
         mBinding.seekBarRange.setOnSeekBarChangeListener(new SeekBar.OnSeekBarChangeListener() {
             @Override
             public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
-                GeoPoint location = new GeoPoint(46.519799, 6.569343);
-                mViewModel.getNewEvents(location, progress).observe(getViewLifecycleOwner(), events -> {
+                mBinding.seekBarValue.setText(String.format(getResources().getConfiguration().locale, "%dkm", progress));
+
+                Location location = mLocationService.getLastKnownLocation(getContext());
+                GeoPoint geoPoint = new GeoPoint(location.getLatitude(), location.getLongitude());
+                mViewModel.getNewEvents(geoPoint, progress).observe(getViewLifecycleOwner(), events -> {
                     mArrayAdapter.clear();
                     mArrayAdapter.addAll(events);
                     mNumberSwipe = 0;
-                    mBinding.seekBarValue.setText(progress +"km");
                 });
             }
